requirements.txt
#data generator
pyyaml
psycopg2-binary
faker

#CDC
python-dotenv
requests

#consumers
kafka-python
boto3
fastparquet
pandas

#dbt
dbt-snowflake